Movies Dataset Correlation Analysis
This Python project performs a comprehensive correlation analysis on a dataset of movies. This goal is to explore and identify relationships between various features such as budget, revenue, genre, ratings, and more. By analyzing these correlations, the project aims to uncover insights that can help in understanding factors that contribute tot a movie's success.