Marathon Races
The Python project involves an in-depth Exploratory Data Analysis (EDA) of a comprehensive dataset on marathon races. The dataset includes various features such as race dates, locations, racers age etc. The primary objective of this analysis is to uncover patterns, trends, and insights that can provide a deeper understanding of factors influencing marathon outcomes.